requirements.txt
django
djangorestframework
markdown
django-filter
django-rest-auth
django-allauth
django-cors-headers
bleach
